+++ core/domain/src/main/java/com/gulderbone/core/domain/pin/Pin.kt
@@ -2,5 +2,5 @@ package com.gulderbone.core.domain.pin
 
 data class Pin(
     val name: String,
-    val value: Int,
+    val value: Long,
 )
