+++ pin/presentation/src/main/java/com/gulderbone/pin/presentation/feature/addpin/AddPinAction.kt
@@ -2,6 +2,6 @@ package com.gulderbone.pin.presentation.feature.addpin
 
 sealed interface AddPinAction {
 
-    // TODO: do not keep as INT, potentially dangerous
-    data class OnAddPinClick(val name: String, val value: Int) : AddPinAction
+    // TODO: do not keep as LONG, potentially dangerous
+    data class OnAddPinClick(val name: String, val value: Long) : AddPinAction
 }
